# lib/api-zod/src/generated/types/appInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AppInfo {
  id: number;
  name: string;
  developer: string;
  description: string;
  shortDescription: string;
  version: string;
  fileSize: string;
  packageName: string;
  category: string;
  tags: string[];
  downloadCount: number;
  minAndroidVersion: string;
  targetAndroidVersion: string;
  permissions: string[];
  whatsNew: string;
  contentRating: string;
  updatedAt: string;
  releasedAt: string;
  downloadUrl: string;
  iconUrl: string;
  averageRating: number;
  totalRatings: number;
}