# lib/api-zod/src/generated/types/rating.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Rating {
  id: number;
  userName: string;
  stars: number;
  reviewText: string;
  createdAt: string;
  helpful: number;
}