# lib/api-zod/src/generated/types/ratingsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Rating } from "./rating";

export interface RatingsResponse {
  ratings: Rating[];
  total: number;
  page: number;
  limit: number;
}