# lib/api-zod/src/generated/types/ratingsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RatingsSummaryBreakdown } from "./ratingsSummaryBreakdown";

export interface RatingsSummary {
  averageRating: number;
  totalRatings: number;
  breakdown: RatingsSummaryBreakdown;
}